+++ tests/FSharp.Markdown.Tests/Markdown.fs
@@ -579,10 +579,10 @@ let ``Transform bulleted lists correctly`` () =
     Markdown.ToHtml doc |> shouldEqual expected
 
 [<Test>]
-let ``Transform header 1 correctly`` () =
+let ``ATX heading without space after hash is treated as paragraph`` () =
     let doc = "#Header 1\nHeader 1\r\n========"
 
-    let expected = "<h1>Header 1</h1>\r\n<h1>Header 1</h1>\r\n" |> properNewLines
+    let expected = "<p>#Header 1</p>\r\n<h1>Header 1</h1>\r\n" |> properNewLines
 
     Markdown.ToHtml doc |> shouldEqual expected
 
@@ -590,7 +590,15 @@ let ``Transform header 1 correctly`` () =
 let ``Transform header 2 correctly`` () =
     let doc = "##Header 2\nHeader 2\r\n--------"
 
-    let expected = "<h2>Header 2</h2>\r\n<h2>Header 2</h2>\r\n" |> properNewLines
+    let expected = "<p>##Header 2</p>\r\n<h2>Header 2</h2>\r\n" |> properNewLines
+
+    Markdown.ToHtml doc |> shouldEqual expected
+
+[<Test>]
+let ``ATX heading with space after hash is transformed to heading element`` () =
+    let doc = "# Header 1\n## Header 2"
+
+    let expected = "<h1>Header 1</h1>\r\n<h2>Header 2</h2>\r\n" |> properNewLines
 
     Markdown.ToHtml doc |> shouldEqual expected